chore: Update project structure
