+++ src/migration/__init__.py
@@ -1,0 +1,204 @@
+from ._version import version as VERSION  # noqa
+
+from .validators import (
+    BaseTableModel,
+    AWSDashboard,
+    Access,
+    AccreditationCertificate,
+    AccreditationSubmission,
+    ActionPlans,
+    AddPlayByPlay,
+    Alias,
+    Annotations,
+    AssessmentsClassroom,
+    AssessmentsDashboard,
+    AssessmentsEvaluation,
+    AssessmentsLister,
+    AssessmentsRating,
+    AssessmentsReport,
+    AssessmentsReportClassroom,
+    AssessmentsSite,
+    Assets,
+    Assignments,
+    CoachingPartnership,
+    CollaboraEndpoint,
+    Comments,
+    ConnectToMerit,
+    ContactForm,
+    CopAssets,
+    CopGroups,
+    CopLocation,
+    CoursesState,
+    CycleDashboard,
+    DiscussionBoards,
+    DiscussionTopics,
+    Documentation,
+    DualLanguageHandbook,
+    DualLanguageSubmission,
+    EaEvalSurvey,
+    Endpoint,
+    FamilySurveyLink,
+    Grade,
+    Inquisitor,
+    InstructorMaterials,
+    Item,
+    Layout,
+    Limit,
+    Links,
+    Lister,
+    Location,
+    Login,
+    Logout,
+    LtiEndpoint,
+    ManageComponents,
+    ManageDatalists,
+    ManageGroups,
+    ManageMenus,
+    ManageRecipes,
+    ManageSite,
+    ManageStudents,
+    ManageAuthUsers,
+    Media,
+    MediaDatabase,
+    MeritEndpoint,
+    MssqlServer,
+    Note,
+    ProgramProfileSubmission,
+    RatingsRubric,
+    RecordsReviewHandbook,
+    RecordsReviewSubmission,
+    ReflectionAssignment,
+    Reports,
+    ReportsBuilder,
+    ResetPassword,
+    ResourceLibrary,
+    SharedAssets,
+    SharedGoals,
+    Solotron,
+    StudentSubmissions,
+    StudentSubmissionsSlug,
+    SupportingDocumentation,
+    Tagit,
+    Tracker,
+    UnassignedAssets,
+    UploadAndReflectionAssignment,
+    UserInvitations,
+    AuthUserSettings,
+    AuthUserSubmissions,
+    AuthUsersLogAllAttributes,
+    AuthUsersLogCompletedGoals,
+    AuthUsersLogCopGroups,
+    AuthUsersLogCopGroupsDigest,
+    AuthUsersLogCreatedAt,
+    AuthUsersLogPartnerships,
+    AuthUsersLogProperties,
+    AuthUsersLogReporting,
+    AuthUsersLogSharedGoals,
+    AuthUsersLogUploads,
+    VideoConference,
+    VideoHighlightSubmission,
+    VideoReflection
+)
+
+__version__ = VERSION
+__author__ = "SSEC"
+
+__all__ = [
+    'BaseTableModel',
+    'AWSDashboard',
+    'Access',
+    'AccreditationCertificate',
+    'AccreditationSubmission',
+    'ActionPlans',
+    'AddPlayByPlay',
+    'Alias',
+    'Annotations',
+    'AssessmentsClassroom',
+    'AssessmentsDashboard',
+    'AssessmentsEvaluation',
+    'AssessmentsLister',
+    'AssessmentsRating',
+    'AssessmentsReport',
+    'AssessmentsReportClassroom',
+    'AssessmentsSite',
+    'Assets',
+    'Assignments',
+    'CoachingPartnership',
+    'CollaboraEndpoint',
+    'Comments',
+    'ConnectToMerit',
+    'ContactForm',
+    'CopAssets',
+    'CopGroups',
+    'CopLocation',
+    'CoursesState',
+    'CycleDashboard',
+    'DiscussionBoards',
+    'DiscussionTopics',
+    'Documentation',
+    'DualLanguageHandbook',
+    'DualLanguageSubmission',
+    'EaEvalSurvey',
+    'Endpoint',
+    'FamilySurveyLink',
+    'Grade',
+    'Inquisitor',
+    'InstructorMaterials',
+    'Item',
+    'Layout',
+    'Limit',
+    'Links',
+    'Lister',
+    'Location',
+    'Login',
+    'Logout',
+    'LtiEndpoint',
+    'ManageComponents',
+    'ManageDatalists',
+    'ManageGroups',
+    'ManageMenus',
+    'ManageRecipes',
+    'ManageSite',
+    'ManageStudents',
+    'ManageAuthUsers',
+    'Media',
+    'MediaDatabase',
+    'MeritEndpoint',
+    'MssqlServer',
+    'Note',
+    'ProgramProfileSubmission',
+    'RatingsRubric',
+    'RecordsReviewHandbook',
+    'RecordsReviewSubmission',
+    'ReflectionAssignment',
+    'Reports',
+    'ReportsBuilder',
+    'ResetPassword',
+    'ResourceLibrary',
+    'SharedAssets',
+    'SharedGoals',
+    'Solotron',
+    'StudentSubmissions',
+    'StudentSubmissionsSlug',
+    'SupportingDocumentation',
+    'Tagit',
+    'Tracker',
+    'UnassignedAssets',
+    'UploadAndReflectionAssignment',
+    'UserInvitations',
+    'AuthUserSettings',
+    'AuthUserSubmissions',
+    'AuthUsersLogAllAttributes',
+    'AuthUsersLogCompletedGoals',
+    'AuthUsersLogCopGroups',
+    'AuthUsersLogCopGroupsDigest',
+    'AuthUsersLogCreatedAt',
+    'AuthUsersLogPartnerships',
+    'AuthUsersLogProperties',
+    'AuthUsersLogReporting',
+    'AuthUsersLogSharedGoals',
+    'AuthUsersLogUploads',
+    'VideoConference',
+    'VideoHighlightSubmission',
+    'VideoReflection'
+]
